- fixed xr::Entity::FindComponent<>();
- fixed xr::Entity::AddComponent<>() - owner is correctly assigned;
- fixed xr::Entity::Clone() - cloning an Entity as a child to itself no longer
  causes infinite recursion, and the owner of cloned Components is now correctly
  assigned;
- fixed xr::StateManager::Pop/Change(), where the current state is now Exit()ed
  first and only Shutdown() once;
- fixed inconsistencies in the uses of int types in the Gfx API;
- added multi-threaded Gfx implementation;
- added support for PointList Primitives in Gfx;

+++ xr3/h/xr/TextureCache.hpp
@@ -125,7 +125,7 @@ private:
   // data
   FormatSpec m_format;
 
-  uint32_t m_size;
+  uint16_t m_size;
   uint32_t m_blockWidth;
   uint32_t m_rowHeight;
